refactor(fieldmap): merge all types into types.ts
